+++ .gitignore
@@ -1,7 +1,6 @@
 .DS_Store
 .dart_tool/
 
-.vscode/
 .packages
 .pub/
 pubspec.lock
